fix ollama create permissions on linux
- give ollama service permissions to read current user's files

+++ scripts/install.sh
@@ -73,8 +73,16 @@ configure_systemd() {
         $SUDO useradd -r -s /bin/false -m -d /usr/share/ollama ollama
     fi
 
-    status "Creating ollama systemd service..."
-    cat <<EOF | $SUDO tee /etc/systemd/system/ollama.service >/dev/null
+    status "Adding current user to ollama group..."
+    $SUDO usermod -a -G ollama $(whoami)
+    # Allow the ollama user to access the current user's files
+    # TODO: This can be removed when model creation does not require access to the current user's files
+    status "Adding ollama to the current user group..."
+    $SUDO usermod -a -G $(whoami) ollama
+
+    if [ ! -f /etc/systemd/system/ollama.service ]; then
+        status "Creating ollama systemd service..."
+        cat <<EOF | $SUDO tee /etc/systemd/system/ollama.service >/dev/null
 [Unit]
 Description=Ollama Service
 After=network-online.target
@@ -91,17 +99,17 @@ Environment="PATH=$PATH"
 [Install]
 WantedBy=default.target
 EOF
-    SYSTEMCTL_RUNNING="$(systemctl is-system-running || true)"
-    case $SYSTEMCTL_RUNNING in
-        running|degraded)
-            status "Enabling and starting ollama service..."
-            $SUDO systemctl daemon-reload
-            $SUDO systemctl enable ollama
-
-            start_service() { $SUDO systemctl restart ollama; }
-            trap start_service EXIT
-            ;;
-    esac
+        $SUDO systemctl daemon-reload
+        $SUDO systemctl enable ollama
+    else
+        SYSTEMCTL_RUNNING="$(systemctl is-system-running || true)"
+        case $SYSTEMCTL_RUNNING in
+            running|degraded)
+                status "Restarting ollama service..."
+                $SUDO systemctl restart ollama
+                ;;
+        esac
+    fi
 }
 
 if available systemctl; then
